pyproject.toml
[project]
name = "debating-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen>=0.9.9",
    "autogen-agentchat>=0.7.4",
    "autogen-ext>=0.7.4",
    "autogen-openai>=0.0.1",
    "fastapi>=0.116.1",
    "groq>=0.31.0",
    "langchain-groq>=0.3.7",
    "openai>=1.101.0",
    "pyautogen>=0.10.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "uvicorn>=0.35.0",
]
